refactor(shell): split server.rs into server/{mod,http,ws,post,log}.rs
- mod.rs: Shared state, frames, start(), timers, broadcast, md5/mime/percent
  utils, malody_root, mma_state writer
- http.rs: 24061 HTTP (static /settings /cover) + WS dispatch
- ws.rs: /ws frame loop (hello/song/state out, diag/control/result in)
- post.rs: 24060 POST + resolve channel (exact-then-substring title match)
- log.rs: log dir logs/, daily-rotated mma-shell-YYYYMMDD.log, readable
  timestamp, level filter
- zero warnings build; E2E + resolve verified after split
